+++ R/print.extrafrail.R
@@ -104,6 +104,23 @@ print.extrafrail<-function (x, digits = max(3L, getOption("digits") - 3L), ...)
                   "pe", "piecewise exponential", x$dist)), " survival function\n", 
                 sep = "")
     }
+    if (x$dist.frail == "MBS") {
+        if (uni) 
+            cat("Univariate mixture of Birnbaum-Saunders frailty model with\n", 
+                ifelse(x$dist == "np", "non-parametric", ifelse(x$dist == 
+                  "pe", "piecewise exponential", x$dist)), " survival function\n", 
+                sep = "")
+        if (bi) 
+            cat("Bivariate mixture of Birnbaum-Saunders frailty model with\n", 
+                ifelse(x$dist == "np", "non-parametric", ifelse(x$dist == 
+                  "pe", "piecewise exponential", x$dist)), " survival function\n", 
+                sep = "")
+        if (!uni & !bi) 
+            cat("Multivariate mixture of Birnbaum-Saunders frailty model with\n", 
+                ifelse(x$dist == "np", "non-parametric", ifelse(x$dist == 
+                  "pe", "piecewise exponential", x$dist)), " survival function\n", 
+                sep = "")
+    }
     cat("\n")
     cat("Coefficients:\n")
     print.default(format(x$coefficients, digits = digits), print.gap = 2L,