+++ R/summary.extrafrail.R
@@ -157,6 +157,23 @@ summary.extrafrail <- function (object, ...)
                   ifelse(object$dist == "pe", "piecewise exponential", 
                     object$dist)), " survival function\n", sep = "")
     }
+    if (object$dist.frail == "MBS") {
+        if (uni) 
+            cat("Univariate mixture of Birnbaum-Saunders frailty model with\n", 
+                ifelse(object$dist == "np", "non-parametric", 
+                  ifelse(object$dist == "pe", "piecewise exponential", 
+                    object$dist)), " survival function\n", sep = "")
+        if (bi) 
+            cat("Bivariate mixture of Birnbaum-Saunders frailty model with\n", 
+                ifelse(object$dist == "np", "non-parametric", 
+                  ifelse(object$dist == "pe", "piecewise exponential", 
+                    object$dist)), " survival function\n", sep = "")
+        if (!uni & !bi) 
+            cat("Multivariate mixture of Birnbaum-Saunders frailty model with\n", 
+                ifelse(object$dist == "np", "non-parametric", 
+                  ifelse(object$dist == "pe", "piecewise exponential", 
+                    object$dist)), " survival function\n", sep = "")
+    }
     if (length(object$coefficients) > 1) {
         if (object$dist == "np") {
             cat("-------------------------------------------------------------------------\n")